feat: refactored company dashboard
