+++ src/pages/CompanyDashboard.jsx
@@ -28,6 +28,7 @@ import {
   TrendingUp,
   Eye,
 } from "lucide-react";
+import AddTask from "../components/CompanyDashboard/AddTask";
 
 const AdminDashboard = () => {
   const [activeSection, setActiveSection] = useState("overview");
@@ -306,155 +307,6 @@ const AdminDashboard = () => {
     </div>
   );
 
-  const renderAddTaskForm = () => {
-    const [title, setTitle] = useState("");
-    const [description, setDescription] = useState("");
-    const [deadline, setDeadline] = useState("");
-    const [tags, setTags] = useState([]);
-    const [tagInput, setTagInput] = useState("");
-    const [assignedInterns, setAssignedInterns] = useState([]);
-
-    const handleAddTag = (e) => {
-      e.preventDefault();
-      if (tagInput.trim() && !tags.includes(tagInput.trim())) {
-        setTags([...tags, tagInput.trim()]);
-        setTagInput("");
-      }
-    };
-
-    const handleRemoveTag = (tagToRemove) => {
-      setTags(tags.filter((tag) => tag !== tagToRemove));
-    };
-
-    const handleSubmit = (e) => {
-      e.preventDefault();
-      const newTask = {
-        title,
-        description,
-        deadline,
-        tags,
-        assignedInterns,
-      };
-      console.log("Task Submitted:", newTask);
-      // You can POST this data to your backend here.
-    };
-
-    return (
-      <div className="max-w-3xl mx-auto space-y-6">
-        <h2 className="text-2xl font-bold text-white">Add New Task</h2>
-
-        <form
-          onSubmit={handleSubmit}
-          className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6 space-y-6"
-        >
-          <div>
-            <label className="block text-gray-300 mb-2">Task Title</label>
-            <input
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              required
-              className="w-full bg-gray-900 border border-gray-700 text-white p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-              placeholder="Enter task title"
-            />
-          </div>
-
-          <div>
-            <label className="block text-gray-300 mb-2">Task Description</label>
-            <textarea
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-              rows={4}
-              required
-              className="w-full bg-gray-900 border border-gray-700 text-white p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-              placeholder="Enter task description"
-            ></textarea>
-          </div>
-
-          <div>
-            <label className="block text-gray-300 mb-2">Deadline</label>
-            <input
-              type="date"
-              value={deadline}
-              onChange={(e) => setDeadline(e.target.value)}
-              required
-              className="w-full bg-gray-900 border border-gray-700 text-white p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-            />
-          </div>
-
-          <div>
-            <label className="block text-gray-300 mb-2">
-              Category / Skill Tags
-            </label>
-            <div className="flex flex-wrap gap-2 mb-2">
-              {tags.map((tag, index) => (
-                <span
-                  key={index}
-                  className="bg-blue-500/20 text-blue-400 px-3 py-1 rounded-full text-sm flex items-center space-x-2"
-                >
-                  {tag}
-                  <button
-                    type="button"
-                    onClick={() => handleRemoveTag(tag)}
-                    className="ml-2 text-red-400 hover:text-red-500"
-                  >
-                    &times;
-                  </button>
-                </span>
-              ))}
-            </div>
-            <div className="flex items-center gap-2">
-              <input
-                type="text"
-                value={tagInput}
-                onChange={(e) => setTagInput(e.target.value)}
-                className="flex-1 bg-gray-900 border border-gray-700 text-white p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-                placeholder="Add a tag"
-              />
-              <button
-                onClick={handleAddTag}
-                className="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 transition"
-              >
-                Add
-              </button>
-            </div>
-          </div>
-
-          <div>
-            <label className="block text-gray-300 mb-2">
-              Assign to Intern(s)
-            </label>
-            <select
-              multiple
-              value={assignedInterns}
-              onChange={(e) =>
-                setAssignedInterns(
-                  Array.from(e.target.selectedOptions, (option) => option.value)
-                )
-              }
-              className="w-full bg-gray-900 border border-gray-700 text-white p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              {interns.map((intern) => (
-                <option key={intern.id} value={intern.name}>
-                  {intern.name} ({intern.department})
-                </option>
-              ))}
-            </select>
-          </div>
-
-          <div className="text-right">
-            <button
-              type="submit"
-              className="bg-gradient-to-r from-blue-500 to-purple-600 px-6 py-3 rounded-lg text-white hover:from-blue-600 hover:to-purple-700 transition-all"
-            >
-              Submit Task
-            </button>
-          </div>
-        </form>
-      </div>
-    );
-  };
-
   const renderTaskList = () => (
     <div className="space-y-8">
       {taskCategories.map((category) => {
@@ -670,513 +522,18 @@ const AdminDashboard = () => {
     </div>
   );
 
-  const renderAddTask = () => {
-    const [formData, setFormData] = useState({
-      title: "",
-      description: "",
-      deliverables: [""],
-      deadline: "",
-      category: "",
-      difficulty: "Intermediate",
-      skillsTags: [],
-      assignedInterns: [],
-      priority: "Medium",
-      estimatedTime: "",
-    });
-
-    const [newSkill, setNewSkill] = useState("");
-    const [showSkillInput, setShowSkillInput] = useState(false);
-
-    const categories = [
-      { id: "development", name: "Development & Programming", icon: Code },
-      { id: "marketing", name: "Marketing & Social Media", icon: Share2 },
-      { id: "design", name: "Design & UI/UX", icon: Target },
-      { id: "content", name: "Content Creation", icon: FileText },
-      { id: "research", name: "Research & Analysis", icon: Brain },
-      { id: "other", name: "Other", icon: MoreHorizontal },
-    ];
-
-    const difficultyLevels = ["Beginner", "Intermediate", "Advanced"];
-    const priorityLevels = ["Low", "Medium", "High", "Critical"];
-
-    const availableSkills = [
-      "JavaScript",
-      "React",
-      "Python",
-      "Node.js",
-      "HTML/CSS",
-      "SQL",
-      "MongoDB",
-      "Social Media Marketing",
-      "Content Writing",
-      "SEO",
-      "Graphic Design",
-      "Data Analysis",
-      "Project Management",
-      "Communication",
-      "Research",
-    ];
-
-    const handleInputChange = (field, value) => {
-      setFormData((prev) => ({
-        ...prev,
-        [field]: value,
-      }));
-    };
-
-    const handleDeliverableChange = (index, value) => {
-      const newDeliverables = [...formData.deliverables];
-      newDeliverables[index] = value;
-      setFormData((prev) => ({
-        ...prev,
-        deliverables: newDeliverables,
-      }));
-    };
-
-    const addDeliverable = () => {
-      setFormData((prev) => ({
-        ...prev,
-        deliverables: [...prev.deliverables, ""],
-      }));
-    };
-
-    const removeDeliverable = (index) => {
-      const newDeliverables = formData.deliverables.filter(
-        (_, i) => i !== index
-      );
-      setFormData((prev) => ({
-        ...prev,
-        deliverables: newDeliverables.length > 0 ? newDeliverables : [""],
-      }));
-    };
-
-    const addSkill = (skill) => {
-      if (skill && !formData.skillsTags.includes(skill)) {
-        setFormData((prev) => ({
-          ...prev,
-          skillsTags: [...prev.skillsTags, skill],
-        }));
-      }
-      setNewSkill("");
-      setShowSkillInput(false);
-    };
-
-    const removeSkill = (skillToRemove) => {
-      setFormData((prev) => ({
-        ...prev,
-        skillsTags: prev.skillsTags.filter((skill) => skill !== skillToRemove),
-      }));
-    };
-
-    const toggleInternAssignment = (internId) => {
-      setFormData((prev) => ({
-        ...prev,
-        assignedInterns: prev.assignedInterns.includes(internId)
-          ? prev.assignedInterns.filter((id) => id !== internId)
-          : [...prev.assignedInterns, internId],
-      }));
-    };
-
-    const handleSubmit = (e) => {
-      e.preventDefault();
-      // Here you would typically send the data to your backend
-      console.log("Task Data:", formData);
-      alert("Task created successfully!");
-    };
-
-    const selectedCategory = categories.find(
-      (cat) => cat.id === formData.category
-    );
-
-    return (
-      <div className="max-w-4xl mx-auto space-y-8">
-        <div className="text-center mb-8">
-          <h2 className="text-3xl font-bold text-white mb-2">
-            Create New Task
-          </h2>
-          <p className="text-gray-400">
-            Fill out the details below to create a new task for your interns
-          </p>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-8">
-          {/* Task Title */}
-          <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-            <label className="block text-white font-semibold mb-3">
-              Task Title <span className="text-red-400">*</span>
-            </label>
-            <input
-              type="text"
-              value={formData.title}
-              onChange={(e) => handleInputChange("title", e.target.value)}
-              className="w-full bg-gray-700/50 border border-gray-600 rounded-lg px-4 py-3 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20"
-              placeholder="Enter a clear, descriptive task title"
-              required
-            />
-          </div>
-
-          {/* Task Description */}
-          <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-            <label className="block text-white font-semibold mb-3">
-              Task Description <span className="text-red-400">*</span>
-            </label>
-            <textarea
-              value={formData.description}
-              onChange={(e) => handleInputChange("description", e.target.value)}
-              rows="4"
-              className="w-full bg-gray-700/50 border border-gray-600 rounded-lg px-4 py-3 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20"
-              placeholder="Provide a detailed description of what needs to be accomplished"
-              required
-            />
-          </div>
-
-          {/* Deliverables */}
-          <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-            <label className="block text-white font-semibold mb-3">
-              Deliverables <span className="text-red-400">*</span>
-            </label>
-            <div className="space-y-3">
-              {formData.deliverables.map((deliverable, index) => (
-                <div key={index} className="flex items-center space-x-3">
-                  <div className="flex-1 flex items-center space-x-3">
-                    <div className="w-6 h-6 bg-blue-500/20 rounded-full flex items-center justify-center">
-                      <span className="text-blue-400 text-sm">{index + 1}</span>
-                    </div>
-                    <input
-                      type="text"
-                      value={deliverable}
-                      onChange={(e) =>
-                        handleDeliverableChange(index, e.target.value)
-                      }
-                      className="flex-1 bg-gray-700/50 border border-gray-600 rounded-lg px-4 py-2 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20"
-                      placeholder="Describe what should be delivered"
-                      required
-                    />
-                  </div>
-                  {formData.deliverables.length > 1 && (
-                    <button
-                      type="button"
-                      onClick={() => removeDeliverable(index)}
-                      className="text-red-400 hover:text-red-300 transition-colors"
-                    >
-                      <svg
-                        className="w-5 h-5"
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M6 18L18 6M6 6l12 12"
-                        />
-                      </svg>
-                    </button>
-                  )}
-                </div>
-              ))}
-              <button
-                type="button"
-                onClick={addDeliverable}
-                className="flex items-center space-x-2 text-blue-400 hover:text-blue-300 transition-colors"
-              >
-                <Plus size={16} />
-                <span>Add Another Deliverable</span>
-              </button>
-            </div>
-          </div>
-
-          {/* Category & Settings Row */}
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-            {/* Category */}
-            <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-              <label className="block text-white font-semibold mb-3">
-                Category <span className="text-red-400">*</span>
-              </label>
-              <div className="space-y-2">
-                {categories.map((category) => {
-                  const Icon = category.icon;
-                  return (
-                    <button
-                      key={category.id}
-                      type="button"
-                      onClick={() => handleInputChange("category", category.id)}
-                      className={`w-full flex items-center space-x-3 px-4 py-3 rounded-lg text-left transition-all duration-200 ${
-                        formData.category === category.id
-                          ? "bg-gradient-to-r from-blue-500/20 to-purple-500/20 text-white border border-blue-500/30"
-                          : "text-gray-300 hover:bg-gray-700/50 hover:text-white border border-gray-600"
-                      }`}
-                    >
-                      <Icon size={18} />
-                      <span>{category.name}</span>
-                    </button>
-                  );
-                })}
-              </div>
-            </div>
-
-            {/* Task Settings */}
-            <div className="space-y-6">
-              {/* Difficulty */}
-              <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-                <label className="block text-white font-semibold mb-3">
-                  Difficulty Level
-                </label>
-                <div className="flex space-x-2">
-                  {difficultyLevels.map((level) => (
-                    <button
-                      key={level}
-                      type="button"
-                      onClick={() => handleInputChange("difficulty", level)}
-                      className={`px-4 py-2 rounded-lg text-sm transition-all duration-200 ${
-                        formData.difficulty === level
-                          ? level === "Beginner"
-                            ? "bg-green-500/20 text-green-400 border border-green-500/30"
-                            : level === "Intermediate"
-                            ? "bg-yellow-500/20 text-yellow-400 border border-yellow-500/30"
-                            : "bg-red-500/20 text-red-400 border border-red-500/30"
-                          : "bg-gray-700/50 text-gray-300 hover:bg-gray-600/50 border border-gray-600"
-                      }`}
-                    >
-                      {level}
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              {/* Priority */}
-              <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-                <label className="block text-white font-semibold mb-3">
-                  Priority
-                </label>
-                <div className="flex space-x-2">
-                  {priorityLevels.map((priority) => (
-                    <button
-                      key={priority}
-                      type="button"
-                      onClick={() => handleInputChange("priority", priority)}
-                      className={`px-4 py-2 rounded-lg text-sm transition-all duration-200 ${
-                        formData.priority === priority
-                          ? priority === "Low"
-                            ? "bg-blue-500/20 text-blue-400 border border-blue-500/30"
-                            : priority === "Medium"
-                            ? "bg-yellow-500/20 text-yellow-400 border border-yellow-500/30"
-                            : priority === "High"
-                            ? "bg-orange-500/20 text-orange-400 border border-orange-500/30"
-                            : "bg-red-500/20 text-red-400 border border-red-500/30"
-                          : "bg-gray-700/50 text-gray-300 hover:bg-gray-600/50 border border-gray-600"
-                      }`}
-                    >
-                      {priority}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {/* Deadline & Estimated Time */}
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-            <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-              <label className="block text-white font-semibold mb-3">
-                Deadline <span className="text-red-400">*</span>
-              </label>
-              <input
-                type="datetime-local"
-                value={formData.deadline}
-                onChange={(e) => handleInputChange("deadline", e.target.value)}
-                className="w-full bg-gray-700/50 border border-gray-600 rounded-lg px-4 py-3 text-white focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20"
-                required
-              />
-            </div>
-
-            <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-              <label className="block text-white font-semibold mb-3">
-                Estimated Time
-              </label>
-              <input
-                type="text"
-                value={formData.estimatedTime}
-                onChange={(e) =>
-                  handleInputChange("estimatedTime", e.target.value)
-                }
-                className="w-full bg-gray-700/50 border border-gray-600 rounded-lg px-4 py-3 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-500/20"
-                placeholder="e.g., 2 weeks, 3 days, 40 hours"
-              />
-            </div>
-          </div>
-
-          {/* Skills Tags */}
-          <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-            <label className="block text-white font-semibold mb-3">
-              Required Skills
-            </label>
-
-            {/* Current Skills */}
-            <div className="flex flex-wrap gap-2 mb-4">
-              {formData.skillsTags.map((skill, index) => (
-                <span
-                  key={index}
-                  className="bg-blue-500/20 text-blue-400 px-3 py-1 rounded-full text-sm flex items-center space-x-2"
-                >
-                  <span>{skill}</span>
-                  <button
-                    type="button"
-                    onClick={() => removeSkill(skill)}
-                    className="text-blue-400 hover:text-blue-300"
-                  >
-                    <svg
-                      className="w-3 h-3"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M6 18L18 6M6 6l12 12"
-                      />
-                    </svg>
-                  </button>
-                </span>
-              ))}
-            </div>
-
-            {/* Available Skills */}
-            <div className="space-y-3">
-              <p className="text-gray-400 text-sm">
-                Select from common skills:
-              </p>
-              <div className="flex flex-wrap gap-2">
-                {availableSkills
-                  .filter((skill) => !formData.skillsTags.includes(skill))
-                  .map((skill) => (
-                    <button
-                      key={skill}
-                      type="button"
-                      onClick={() => addSkill(skill)}
-                      className="bg-gray-700/50 text-gray-300 px-3 py-1 rounded-full text-sm hover:bg-gray-600/50 transition-colors"
-                    >
-                      + {skill}
-                    </button>
-                  ))}
-              </div>
-
-              {/* Custom Skill Input */}
-              <div className="flex items-center space-x-2">
-                {showSkillInput ? (
-                  <div className="flex items-center space-x-2">
-                    <input
-                      type="text"
-                      value={newSkill}
-                      onChange={(e) => setNewSkill(e.target.value)}
-                      className="bg-gray-700/50 border border-gray-600 rounded-lg px-3 py-2 text-white placeholder-gray-400 focus:outline-none focus:border-blue-500"
-                      placeholder="Enter custom skill"
-                      onKeyPress={(e) =>
-                        e.key === "Enter" &&
-                        (e.preventDefault(), addSkill(newSkill))
-                      }
-                    />
-                    <button
-                      type="button"
-                      onClick={() => addSkill(newSkill)}
-                      className="bg-blue-500/20 text-blue-400 px-3 py-2 rounded-lg hover:bg-blue-500/30 transition-colors"
-                    >
-                      Add
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => (
-                        setShowSkillInput(false), setNewSkill("")
-                      )}
-                      className="text-gray-400 hover:text-gray-300"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                ) : (
-                  <button
-                    type="button"
-                    onClick={() => setShowSkillInput(true)}
-                    className="text-blue-400 hover:text-blue-300 text-sm"
-                  >
-                    + Add Custom Skill
-                  </button>
-                )}
-              </div>
-            </div>
-          </div>
-
-          {/* Assign to Interns */}
-          <div className="bg-gray-800/40 backdrop-blur-lg border border-gray-700/50 rounded-2xl p-6">
-            <label className="block text-white font-semibold mb-3">
-              Assign to Intern(s)
-            </label>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-              {interns.map((intern) => (
-                <button
-                  key={intern.id}
-                  type="button"
-                  onClick={() => toggleInternAssignment(intern.id)}
-                  className={`p-4 rounded-lg border transition-all duration-200 ${
-                    formData.assignedInterns.includes(intern.id)
-                      ? "bg-gradient-to-r from-blue-500/20 to-purple-500/20 border-blue-500/30 text-white"
-                      : "bg-gray-700/50 border-gray-600 text-gray-300 hover:bg-gray-600/50"
-                  }`}
-                >
-                  <div className="flex items-center space-x-3">
-                    <div className="w-10 h-10 bg-gradient-to-r from-blue-500 to-purple-600 rounded-full flex items-center justify-center">
-                      <span className="text-white font-bold text-sm">
-                        {intern.name
-                          .split(" ")
-                          .map((n) => n[0])
-                          .join("")}
-                      </span>
-                    </div>
-                    <div className="text-left">
-                      <p className="font-medium">{intern.name}</p>
-                      <p className="text-sm opacity-75">{intern.department}</p>
-                    </div>
-                  </div>
-                </button>
-              ))}
-            </div>
-          </div>
-
-          {/* Submit Button */}
-          <div className="flex justify-end space-x-4">
-            <button
-              type="button"
-              onClick={() => setActiveSection("task-list")}
-              className="px-6 py-3 bg-gray-700 text-white rounded-lg hover:bg-gray-600 transition-colors"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="px-6 py-3 bg-gradient-to-r from-blue-500 to-purple-600 text-white rounded-lg hover:from-blue-600 hover:to-purple-700 transition-all duration-300 font-medium"
-            >
-              Create Task
-            </button>
-          </div>
-        </form>
-      </div>
-    );
-  };
-
   const renderContent = () => {
     switch (activeSection) {
       case "overview":
         return renderOverview();
       case "add-task":
-        return (
-          <div className="text-center py-20">
-            <h2 className="text-2xl font-bold text-white mb-4">Add New Task</h2>
-            <p className="text-gray-400">Task creation form would go here</p>
-          </div>
-        );
+        return <AddTask />;
+      // return (
+      //   <div className="text-center py-20">
+      //     <h2 className="text-2xl font-bold text-white mb-4">Add New Task</h2>
+      //     <p className="text-gray-400">Task creation form would go here</p>
+      //   </div>
+      // );
       case "task-list":
         return renderTaskList();
       case "interns-list":
